AFNI HRF Specifications

The `AFNI_HRF` function creates an object representing an AFNI-specific hemodynamic response function (HRF). It is a class constructor for AFNI HRFs.

Usage

AFNI_HRF(name, nbasis, params, span = 24)

Arguments

name

A string specifying the name of the AFNI HRF.

nbasis

An integer representing the number of basis functions for the AFNI HRF.

params

A list containing the parameter values for the AFNI HRF.

span

A numeric value representing the span in seconds of the HRF. Default is 24.

Value

An AFNI_HRF object with the specified properties.

Details

This file contains all AFNI HRF (Hemodynamic Response Function) specifications, constructors, and related functions for creating AFNI-compatible HRF models.

AFNI HRF Constructor Function

See also